Fix `ObsmetaType` to account for incomplete taxonomies
This change allows incomplete taxonomies to be present in the
observation metadata.
Incomplete taxonomies are defined as taxonomies data that does not
contain all levels down to `species`.
Hence, if a `biom` file contains just ['kingdom', 'phylum'] it is still
valid.
This change also allows for empty data ('') to be present in a taxonomy
name.