Fix: Hashing module inputs updated to support collect operator
The earlier implementation with multiple outputs in one channel does not
work well with the `collect` operator. The solution was to create
multiple channels, each with a single output and then use `collect`
independently on each one. We do not care about the ordering here since
we do not change the file names and those are what the process uses internally